move stuff around to generally work
postgres wants the db opened for each request and closed after
bolt wants it opened once at server startup and closed on server shutdown

this will take some work
