+++ impl/pkg/storage/db/bolt/bolt_test.go
@@ -1,4 +1,4 @@
-package storage
+package bolt
 
 import (
 	"os"
